+++ command/commit.js
@@ -50,7 +50,7 @@ if (!found) {
 }
 logger.info('node_modules ignored.');
 
-runner('git add .gitignore && git add .', OPTIONS.path); 
+runner('git add .gitignore && git add .', OPTIONS.path);
 runner('git commit -m "Auto committed by yuan-npm-release"', OPTIONS.path);
 logger.info('Committed to local repository.');